python-service/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
openai==0.28.0
python-dotenv==1.0.0
pydantic==2.4.2
starlette==0.27.0
typing-extensions==4.8.0 